requirements.txt
asgiref==3.5.2
coverage==6.4.2
Django==4.0.6
django-extensions==3.2.0
djangorestframework==3.13.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
PyJWT==2.4.0
python-decouple==3.6
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.1